Cleans up some compiler warnings due to type formatting issues.

+++ hfaxd/Status.c++
@@ -1,4 +1,4 @@
-/*	$Id: Status.c++ 1068 2011-11-29 06:22:42Z faxguy $ */
+/*	$Id: Status.c++ 1120 2012-09-24 17:19:11Z faxguy $ */
 /*
  * Copyright (c) 1995-1996 Sam Leffler
  * Copyright (c) 1995-1996 Silicon Graphics, Inc.
@@ -248,7 +248,7 @@ HylaFAXServer::Mprintf(FILE* fd, const char* fmt, const ModemConfig& config)
 		if (c == '%')		// %% -> %
 		    putc(c, fd);
 		else
-		    fprintf(fd, "%.*s%c", fp-fspec, fspec, c);
+		    fprintf(fd, "%.*s%c", (int) (fp-fspec), fspec, c);
 		continue;
 	    }
 	    fp[0] = mformat[c-'a'];	// printf format string
